Code clean up, testing clean up

+++ App/city-feedback/src/main/java/com/example/city_feedback/complaintManagement/infrastructure/repositories/LocationRepository.java
@@ -6,10 +6,22 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
+/**
+ * Repository interface for managing Location entities.
+ * Provides methods for accessing and managing Location data.
+ */
 @Repository
-public interface LocationRepository extends JpaRepository<Location, Long> {
+public interface LocationRepository extends JpaRepository<Location, Integer> {
 
+    /**
+     * Finds a location by its street, house number, postal code, and city.
+     *
+     * @param street      the street name
+     * @param houseNumber the house number
+     * @param postalCode  the postal code
+     * @param city        the city
+     * @return an optional Location matching the provided address details
+     */
     Optional<Location> findByStreetAndHouseNumberAndPostalCodeAndCity(
-            String street, String houseNumber, String postalCode, String city
-    );
+            String street, String houseNumber, String postalCode, String city);
 }
